Functional implementation of caching results
This creates a cache of json files in the SAISTCache folder storing any
and all findings per file, allowing SAIST to skip sending excess llm
requests if the file has been analyzed before.

It checks the hash of the file contents to verify that the file hasn't
been altered since the cache of findings was made, and if it hasn't,
it'll pull the findings from the cache, instead of requerying the llm.

This will massively speed up reruns of SAIST, and even allow for SAIST
to continue from where it was if the program is closed prematurely.

The implementation in this commit should work for all scms however,
the code needs cleaned up a bit, and tested further.