Implement result caching (#47)
* Functional implementation of caching results

This creates a cache of json files in the SAISTCache folder storing any
and all findings per file, allowing SAIST to skip sending excess llm
requests if the file has been analyzed before.

It checks the hash of the file contents to verify that the file hasn't
been altered since the cache of findings was made, and if it hasn't,
it'll pull the findings from the cache, instead of requerying the llm.

This will massively speed up reruns of SAIST, and even allow for SAIST
to continue from where it was if the program is closed prematurely.

The implementation in this commit should work for all scms however,
the code needs cleaned up a bit, and tested further.

* Add --disable-caching flag

This flag allows the end user to turn off the caching, meaning every
file will be analyzed by an llm, regardless as to whether or not a
cache folder exists already.

Also cleaned up the code a little

* General tidying up

Cleaned up some more of the code changes

* Add --cache-folder argument

This allows the end user to specify a foldername for the caching

* Clean up code a little

* Fix security issue

* Set encoding to UTF-8 on cache files

* Change from md5 to sha256 to reduce cache collisions

* Update README.md with new caching flags

* Fix docker example